fix: esp使用串口2, pwm使用时钟2

+++ User/Inc/pwm.h
@@ -1,4 +1,4 @@
 #include "main.h"
 
 void pwm_init(void);
-void set_pwm(uint8_t timer, uint8_t channel, uint8_t percent);
+void set_pwm(uint8_t timer, uint8_t channel, float percent);
